+++ kernel.c
@@ -67,11 +67,11 @@ void kmain(multiboot_info_t *multiboot_info)
 	while(1);
 }
 
-void interrupt_handler(u32 cr2, u32 edi, u32 esi, u32 ebp, u32 esp, u32 ebx, u32 edx, u32 ecx, u32 eax, const u32 interrupt_no, u32 error_code, u32 eip) {
+void interrupt_handler(u32 cr2, u32 edi, u32 esi, u32 ebp, u32 esp, u32 ebx, u32 edx, u32 ecx, u32 eax, const u32 interrupt_no, u32 error_code, size_t eip) {
 	if(interrupt_no < 32) {
-		kpanic_fmt("Exception %d (%s) at 0x%x, error %d\n", (u64) interrupt_no, exceptions_string[interrupt_no], (u64) eip, (u64) error_code);
+		kpanic_fmt("Exception %d (%s) at 0x%x, error %d\n", (u32) interrupt_no, exceptions_string[interrupt_no], (size_t) eip, error_code);
 	} else {
-		kpanic_fmt("Interrupt %d (%s) at 0x%x, error %d\n", (u64) interrupt_no - 32, interrupts_string[interrupt_no - 32], (u64) eip, (u64) error_code);
+		kpanic_fmt("Interrupt %d (%s) at 0x%x, error %d\n", (u32) interrupt_no - 32, interrupts_string[interrupt_no - 32], (size_t) eip, error_code);
 	}
 
 	if(interrupt_no == 0) { //Don't know what to do yet so just ignore
